requirements.txt
Django==1.11.25
djangorestframework==3.9.4
honcho==1.0.1
psycopg2==2.8.4
pytz==2019.3
